Update ths-shell-init.sh
bs

+++ bootstrap/ths-shell-init.sh
@@ -179,11 +179,9 @@ e=$? && handleError
 
 echo "[ths-shell-init] Create bin folder ..."
 tstep=1
-echo "mkdir -p /ths/$thsStorage/bin"
 mkdir -p /ths/$thsStorage/bin
 e=$? && handleError
-echo "==="
-ln -sf /ths/$thsStorage/bin /ths/bin
+if [ ! -e /ths/bin ]; then ln -sf /ths/$thsStorage/bin /ths/bin; fi
 e=$? && handleError 
 cd /ths/bin && ln -sf /ths/tinyHomeServer/bin/jupyter-pass jupyter-pass
 e=$? && handleError
